created post and search folder for both pages to make files more organize
