+++ imports/startup/client/routes.jsx
@@ -2,8 +2,8 @@ import { Meteor } from 'meteor/meteor';
 import React from 'react';
 import { Router, Route, browserHistory} from 'react-router';
 import Homepage from '/imports/ui/pages/home/Homepage.jsx';
-import SearchPage from '/imports/ui/pages/home/SearchPage.jsx';
-import JobPostPage from '/imports/ui/pages/home/JobPostPage.jsx';
+import SearchPage from '/imports/ui/pages/search/SearchPage.jsx';
+import JobPostPage from '/imports/ui/pages/post/JobPostPage.jsx';
 import GoogleMapsPage from '/imports/ui/pages/home/GoogleMapsPage.jsx';
 
 export default () =>(
